// src/interfaces/UserType.ts
export type UserType = {
  about_me: string;
  account_id: number;
  answer_count: number;
  creation_date: number;
  display_name: string;
  last_access_date: number;
  location: string;
  profile_image: string;
  question_count: number;
  reputation: number;
  user_id: number;
  view_count: number;
  website_url: string;
};